requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysqlclient==2.2.4
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.0
twilio==9.2.3
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
